base_802_11_defs: management frame type enum to be unsigned char
ap_wlan_hal_dwpal parses msg_type as part of STA_SOFTBLOCK_DROP
event handling, while dwpal parser returns unsigned char value
but comparision is done against int value of management frame type.
ie:
message_type == s80211MgmtFrame::eType::AUTH

Fix by forcing enum managament frame type to be unsigned char.

PPM-1866

+++ common/beerocks/bwl/dwpal/ap_wlan_hal_dwpal.cpp
@@ -2691,7 +2691,7 @@ bool ap_wlan_hal_dwpal::process_dwpal_event(char *buffer, int bufLen, const std:
 
         /* TEMP: Traces... */
         LOG(DEBUG) << "numOfValidArgs[1]=" << numOfValidArgsForMsgType[1]
-                   << ", message_type=" << (int)message_type;
+                   << ", message_type=" << (uint8_t)message_type;
 
         if (message_type == s80211MgmtFrame::eType::PROBE_REQ) {
